Div Galore
I added divs to simulate pixels and used the 'difference' property as a safeguard to protect the text from the array of pixels. Currently, I'm troubleshooting how to ensure the divs always fill the screen. I think there's a way to do it using P5.js, but I'm unsure if that counts as JavaScript. We might end up adding more pixels to our HTML, but the space at the end of the body might look a little off.