trouble shooting and planning for the next t
we deleted z index to make the bottom text visible. Now we have to style our mobile designs. We are going to add titles to some of the paragraphs in order to make a table of content. We might also add in a scroll to effect using href with the table of content. Sand by with type experimentation until we finish mobile.