Applied feedback
Made the leading more condensed for mobile. The rag looks a little better. Added a new style for the navs. The edges will be a grid, and then we can color them to create a pixelated effect. Also added a little bit of space at the top of each nav to improve the flow. Currently working on troubleshooting.